[ACLRequests strong typing] moved class `MailboxName` from inside `imap/message/request/SetACLRequest` to `imap/message/` for common usage
