+++ derby.log
@@ -1,0 +1,13 @@
+----------------------------------------------------------------
+Thu Sep 26 16:10:20 CEST 2024:
+Booting Derby version The Apache Software Foundation - Apache Derby - 10.14.2.0 - (1828579): instance a816c00e-0192-2eab-240b-0000037739a8 
+on database directory memory:/home/fazavant/idea-projects/james-project/mailboxintegrationtesting with class loader jdk.internal.loader.ClassLoaders$AppClassLoader@4e0e2f2a 
+Loaded from file:/home/fazavant/.m2/repository/org/apache/derby/derby/10.14.2.0/derby-10.14.2.0.jar
+java.vendor=Oracle Corporation
+java.runtime.version=21.0.3+7-LTS-152
+user.dir=/home/fazavant/idea-projects/james-project
+os.name=Linux
+os.arch=amd64
+os.version=6.8.0-45-generic
+derby.system.home=null
+Database Class Loader started - derby.database.classpath=''
